👩‍🔬 Add cite nodes to myst-spec-ext
