+++ .changeset/serious-ducks-sip.md
@@ -1,0 +1,9 @@
+---
+'myst-cli': patch
+'myst-roles': patch
+'myst-spec-ext': patch
+'myst-to-docx': patch
+'tex-to-myst': patch
+---
+
+Add citation nodes to myst-spec-ext
