+++ packages/myst-cli/package.json
@@ -81,6 +81,7 @@
     "myst-frontmatter": "^0.0.9",
     "myst-parser": "^0.0.20",
     "myst-spec": "^0.0.4",
+    "myst-spec-ext": "^0.0.6",
     "myst-templates": "^0.1.10",
     "myst-to-docx": "^0.0.15",
     "myst-to-tex": "^0.0.18",
