+++ packages/tex-to-myst/package.json
@@ -42,6 +42,7 @@
     "@unified-latex/unified-latex": "^1.2.2",
     "myst-common": "^0.0.14",
     "myst-frontmatter": "^0.0.9",
+    "myst-spec-ext": "^0.0.6",
     "unist-util-select": "^4.0.0",
     "vfile-reporter": "^7.0.4"
   },
